Implement focus for addressbar.

+++ src/layout.rs
@@ -112,15 +112,15 @@ impl ComputedLocation {
             ComputedLocation::Computed(loc) => { loc.height },
         }
     }
-    pub fn is_inside(&self, x: u32, y: u32) -> bool {
+    pub fn is_inside(&self, x: f32, y: f32) -> bool {
         //TODO: for now we use this to check pixel values, but we actually should convert units properly somewhere (before the renderer, I guess)
         //      in general we need to do a pass on using correct units everywhere
         return match self {
             ComputedLocation::NotYetComputed => panic!("Node has not yet been computed"),
             ComputedLocation::Computed(loc) => {
-                x as f32 >= loc.x && x as f32 <= loc.x + loc.width
+                x >= loc.x && x <= loc.x + loc.width
                 &&
-                y as f32 >= loc.y && y as f32 <= loc.y + loc.height
+                y >= loc.y && y <= loc.y + loc.height
             },
         }
     }
